requirements.txt
pygetwindow
Pillow
pyautogui
